correcciones MPC

+++ Control 3/TPA/3/codigo/include/Ky.h
@@ -3,6 +3,6 @@
 #ifndef KY_H
 #define KY_H
 
-const double Ky = 3.388893;
+const float Ky = 3.388893e+00;
 
 #endif // KY_H
